fix: Fix Rust scorer to match Python implementation exactly
- Fixed score calculation to return raw weighted score instead of binary score
- Added PBKDF2-SHA256 support for API key validation (Django's default)
- Fixed all database type mismatches (i32 -> i64 for BIGINT fields)
- Added missing trusted issuers to match Python configuration
- Fixed event action codes (SCU/LDP instead of full names)
- Fixed timestamp formatting to match Python's 6 decimal places
- Reordered JSON fields to match Python response exactly
- Added fallback from BinaryWeightedScorer to WeightedScorer table
- Fixed table names (ceramic_cache_ceramiccache not ceramic_cache)
- Removed non-existent columns from Django model queries

The Rust scorer now produces byte-for-byte identical responses to Python
(except for timestamps which differ by request time).

+++ rust-scorer/src/dedup/lifo.rs
@@ -43,7 +43,7 @@ pub struct LifoResult {
 pub async fn lifo_dedup(
     stamps: &[ValidStamp],
     address: &str,
-    community_id: i32,
+    community_id: i64,
     weights: &HashMap<String, Decimal>,
     tx: &mut Transaction<'_, Postgres>,
 ) -> Result<LifoResult, DatabaseError> {
@@ -88,7 +88,7 @@ pub async fn lifo_dedup(
 async fn lifo_dedup_attempt(
     stamps: &[ValidStamp],
     address: &str,
-    community_id: i32,
+    community_id: i64,
     weights: &HashMap<String, Decimal>,
     tx: &mut Transaction<'_, Postgres>,
 ) -> Result<LifoResult, DatabaseError> {
